Cleaned Some things up
